requirements.txt
django
django-webpack-loader
django-debug-toolbar
Pillow
gunicorn
psycopg2-binary
djangorestframework
